Put each model in its own Python file, to reduce the chance of having circular references on imports.
