+++ connect/models/__init__.py
@@ -3,24 +3,55 @@
 # This file is part of the Ingram Micro Cloud Blue Connect SDK.
 # Copyright (c) 2019 Ingram Micro. All Rights Reserved.
 
-from .activation_response import ActivationTemplateResponse, ActivationTileResponse
+from .activation import Activation
+from .activation_template_response import ActivationTemplateResponse
+from .activation_tile_response import ActivationTileResponse
+from .agreement import Agreement
+from .agreement_stats import AgreementStats
 from .asset import Asset
 from .base import BaseModel
-from .company import Company, User
+from .company import Company
+from .configuration import Configuration
 from .connection import Connection
-from .contact import Contact, ContactInfo, PhoneNumber
-from .conversation import Conversation, ConversationMessage
-from .event import EventInfo, Events
+from .constraints import Constraints
+from .contact import Contact
+from .contact_info import ContactInfo
+from .contract import Contract
+from .conversation import Conversation
+from .conversation_message import ConversationMessage
+from .customer_ui_settings import CustomerUiSettings
+from .document import Document
+from .download_link import DownloadLink
+from .event import Event
+from .events import Events
+from .ext_id_hub import ExtIdHub
 from .fulfillment import Fulfillment
-from .hub import Hub, HubInstance, ExtIdHub, HubStats
-from .marketplace import Activation, Agreement, AgreementStats, Contract, Marketplace
-from .parameters import Constraints, Param, ValueChoice
-from .product import CustomerUiSettings, Document, DownloadLink, Item, Product, \
-    ProductCategory, ProductConfiguration, ProductFamily, ProductStats, ProductStatsInfo, Renewal
+from .hub import Hub
+from .hub_instance import HubInstance
+from .hub_stats import HubStats
+from .item import Item
+from .marketplace import Marketplace
+from .param import Param
+from .phone_number import PhoneNumber
+from .product import Product
+from .product_category import ProductCategory
+from .product_configuration import ProductConfiguration
+from .product_family import ProductFamily
+from .product_stats import ProductStats
+from .product_stats_info import ProductStatsInfo
+from .renewal import Renewal
 from .server_error_response import ServerErrorResponse
-from .tier_config import Configuration, Template, TierAccount, TierAccounts, TierConfig, \
-    TierConfigRequest
-from .usage import UsageFile, UsageListing, UsageRecord, UsageRecords
+from .template import Template
+from .tier_account import TierAccount
+from .tier_accounts import TierAccounts
+from .tier_config import TierConfig
+from .tier_config_request import TierConfigRequest
+from .usage_file import UsageFile
+from .usage_listing import UsageListing
+from .usage_record import UsageRecord
+from .usage_records import UsageRecords
+from .user import User
+from .value_choice import ValueChoice
 
 __all__ = [
     'Activation',
@@ -42,7 +73,7 @@ __all__ = [
     'CustomerUiSettings',
     'Document',
     'DownloadLink',
-    'EventInfo',
+    'Event',
     'Events',
     'ExtIdHub',
     'Fulfillment',
