+++ connect/models/customer_ui_settings.py
@@ -1,0 +1,29 @@
+# -*- coding: utf-8 -*-
+
+# This file is part of the Ingram Micro Cloud Blue Connect SDK.
+# Copyright (c) 2019 Ingram Micro. All Rights Reserved.
+
+from typing import List
+
+from .base import BaseModel
+from .document import Document
+from .download_link import DownloadLink
+from .schemas import CustomerUiSettingsSchema
+
+
+class CustomerUiSettings(BaseModel):
+    """ Customer Ui Settings for a product. """
+
+    _schema = CustomerUiSettingsSchema()
+
+    description = None  # type: str
+    """ (str) Description. """
+
+    getting_started = None  # type: str
+    """ (str) Getting started. """
+
+    download_links = None  # type: List[DownloadLink]
+    """ (List[:py:class:`.DownloadLink`]) Download links. """
+
+    documents = None  # type: List[Document]
+    """ (List[:py:class:`.Document`]) Documents. """
