+++ connect/models/document.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+
+# This file is part of the Ingram Micro Cloud Blue Connect SDK.
+# Copyright (c) 2019 Ingram Micro. All Rights Reserved.
+
+from .base import BaseModel
+from .schemas import DocumentSchema
+
+
+class Document(BaseModel):
+    """ Document for a product. """
+
+    _schema = DocumentSchema()
+
+    title = None  # title: str
+    """ (str) Document title. """
+
+    url = None  # title: str
+    """ (str) Document URL. """
