+++ connect/models/renewal.py
@@ -1,0 +1,27 @@
+# -*- coding: utf-8 -*-
+
+# This file is part of the Ingram Micro Cloud Blue Connect SDK.
+# Copyright (c) 2019 Ingram Micro. All Rights Reserved.
+
+import datetime
+
+from .base import BaseModel
+from .schemas import RenewalSchema
+
+
+class Renewal(BaseModel):
+    """ Item renewal data. """
+
+    _schema = RenewalSchema()
+
+    from_ = None  # type: datetime.datetime
+    """ (datetime.datetime) Date of renewal beginning. """
+
+    to = None  # type: datetime.datetime
+    """ (datetime.datetime) Date of renewal end. """
+
+    period_delta = None  # type: int
+    """ (int) Size of renewal period. """
+
+    period_uom = None  # type: str
+    """ (str) Unit of measure for renewal period. One of: year, month, day, hour. """
