+++ connect/models/tier_config.py
@@ -5,79 +5,17 @@
 
 from typing import Optional, List
 
-import connect.models
 from .base import BaseModel
-from .company import User
+from .configuration import Configuration
 from .connection import Connection
-from .contact import ContactInfo
-from .event import Events
-from .marketplace import Activation
-from .parameters import Param
+from .contract import Contract
+from .events import Events
+from .marketplace import Marketplace
+from .param import Param
 from .product import Product
-from connect.models.schemas import TemplateSchema, TierAccountSchema, \
-    TierAccountsSchema, TierConfigSchema, TierConfigRequestSchema, ConfigurationSchema
-
-
-class TierAccount(BaseModel):
-    """ Tier account. """
-
-    _schema = TierAccountSchema()
-
-    name = None  # type: str
-    """ (str) Tier name. """
-
-    contact_info = None  # type: ContactInfo
-    """ (:py:class:`.ContactInfo`) Tier Contact Object. """
-
-    external_id = None  # type: Optional[str]
-    """ (str|None) Only in case of filtering by this field. """
-
-    external_uid = None  # type: Optional[str]
-    """ (str|None) Only in case of filtering by this field. """
-
-
-class TierAccounts(BaseModel):
-    """ TierAccounts object. """
-
-    _schema = TierAccountsSchema()
-
-    customer = None  # type: TierAccount
-    """ (:py:class:`.TierAccount`) Customer Level TierAccount Object. """
-
-    tier1 = None  # type: TierAccount
-    """ (:py:class:`.TierAccount`) Level 1 TierAccount Object. """
-
-    tier2 = None  # type: TierAccount
-    """ (:py:class:`.TierAccount`) Level 2 TierAccount Object. """
-
-
-class Template(BaseModel):
-    """ Tier Template """
-
-    _schema = TemplateSchema()
-
-    name = None  # type: str
-    """ (str) Template name. """
-
-    representation = None  # type: str
-    """ (str) Template representation. """
-
-
-class Configuration(BaseModel):
-    """ Configuration Phase Parameter Context-Bound Data Object.
-
-    To be used in parameter contexts:
-
-    - Asset.
-    - Fulfillment Request.
-    - TierConfig.
-    - TierConfig Requests.
-    """
-
-    _schema = ConfigurationSchema()
-
-    params = None  # type: List[Param]
-    """ (List[:py:class:`.Param`])  """
+from .template import Template
+from .tier_account import TierAccount
+from .schemas import TierConfigSchema
 
 
 class TierConfig(BaseModel):
@@ -113,10 +51,10 @@ class TierConfig(BaseModel):
     template = None  # type: Template
     """ (:py:class:`.Template`) Template Object.  """
 
-    contract = None  # type: connect.models.Contract
+    contract = None  # type: Contract
     """ (:py:class:`.Contract`) Contract Object reference. """
 
-    marketplace = None  # type: connect.models.Marketplace
+    marketplace = None  # type: Marketplace
     """ (:py:class:`.Marketplace`) Marketplace Object reference. """
 
     configuration = None  # type: Configuration
@@ -141,6 +79,7 @@ class TierConfig(BaseModel):
         :return: The requested Tier Config, or ``None`` if it was not found.
         :rtype: Optional[TierConfig]
         """
+        from .tier_config_request import TierConfigRequest
         from connect.resources.base import ApiClient
 
         response, _ = ApiClient(config, base_path='tier/config-requests').get(
@@ -159,71 +98,6 @@ class TierConfig(BaseModel):
 
     def get_param_by_id(self, id_):
         """ Get a Tier Config parameter.
-
-        :param str id_: Parameter id.
-        :return: The requested parameter, or ``None`` if it was not found.
-        :rtype: Param
-        """
-        try:
-            return list(filter(lambda param: param.id == id_, self.params))[0]
-        except IndexError:
-            return None
-
-
-class TierConfigRequest(BaseModel):
-    _schema = TierConfigRequestSchema()
-
-    type = None  # type: str
-    """ (str) TCR type. One of: setup, update. """
-
-    status = None  # type: str
-    """ (str) TCR current status. One of: tiers_setup, pending, inquiring, approved, failed. """
-
-    configuration = None  # type: TierConfig
-    """ (:py:class:`.TierConfig`) Full representation of TierConfig Object. """
-
-    parent_configuration = None  # type: Optional[TierConfig]
-    """ (:py:class:`.TierConfig` | None) Full representation of parent TierConfig. """
-
-    account = None  # type: TierAccount
-    """ (:py:class:`.TierAccount`) Reference object to TierAccount. """
-
-    product = None  # type: Product
-    """ (:py:class:`.Product`) Reference object to product (application). """
-
-    tier_level = None  # type: int
-    """ (int) Tier level for product from customer perspective (1 or 2). """
-
-    params = None  # type: List[Param]
-    """ (List[:py:class:`.Param`]) List of parameter data objects as in Asset Object.
-    Params can be modified only in Pending state.
-    """
-
-    environment = None  # type: str
-    """ (str) TCR environment (test, prod or preview) """
-
-    assignee = None  # type: Optional[User]
-    """ (:py:class:`.User` | None) TCR environment. One of: test, prod, preview. """
-
-    template = None  # type: Optional[Template]
-    """ (:py:class:`.Template` | None) Template Object. This is filled only if TCR is approved. """
-
-    reason = None  # type: Optional[str]
-    """ (str|None) Failing reason. This is filled only if TCR is failed. """
-
-    activation = None  # type: Optional[Activation]
-    """ (:py:class:`.Activation` | None) Activation object. This is created only if TCR
-    has ordering parameters and seen in inquiring state of the TCR.
-    """
-
-    notes = None  # type: Optional[str]
-    """ (str) TCR pending notes. Notes can be modified only in Pending state. """
-
-    events = None  # type: Optional[Events]
-    """ (:py:class:`.Events` | None) Tier Config request Events. """
-
-    def get_param_by_id(self, id_):
-        """ Get a Tier Config Request parameter.
 
         :param str id_: Parameter id.
         :return: The requested parameter, or ``None`` if it was not found.
